requirements.txt
alembic==1.17.2
fastapi==0.128.0
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
sqlalchemy==2.0.45
uvicorn==0.40.0
